getting PDSegments in tables, check 10.1.1.10.1035-Table-2.json

+++ src/main/scala/edu/psu/sagnik/research/table/model/TableADT.scala
@@ -33,8 +33,8 @@ case class IntermediateTable(bb: Rectangle,
                              mention: Option[String],
                              pageNo:Int,
                              pdLines: Option[Seq[PDSegment]],
-                             pageHeight: Int,
-                             pageWidth: Int) //this will be later replaced with PDLines from pdfxtk.
+                             pageHeight: Float,
+                             pageWidth: Float) //this will be later replaced with PDLines from pdfxtk.
 
 case class Table(pageNumber:Int, //page number
   bb: Rectangle,
